Separate ScrollSpy into abstract class
